Scale compression tol with L, implement hard_restart adaptive compression.
Before this commit, the compression tolerance was only scaled with the
HSS level in the new stopping criterion (used in the incremental
'stable' adaptive scheme), but not in the actual rank-revealing QR
call. Now the compression tolerance is always scaled with the HSS
level, both in the stopping criterion and in the actual RRQR, for both
the doubling ('original') and incremental ('stable') strategies. This
is implemented for both the sequential and MPI codes.

A new adaptive compression scheme was also added called
'hard_restart'. If compression fails to reach the prescribed
tolerance, the entire compression is restarted, including for the
nodes of the HSS tree that were already succesfully compressed. (the
random sampling of the first columns is reused). This mimics a
poor-man's adaptive scheme.