Implement a fallback for when the total receive or send buffersize in all_to_all_v overflows the 32-bit int maximum value.
An overflow of the total buffer size likely leads to
an overflow in the displacements). In this case we fall back
to a simple Isend/Irecv loop. This might be slower than
splitting in multiple MPI_Alltoallv calls, but it also
avoids a copy from the input buffer.

An overflow can still occur if an individual message is too large.
But in that case, one could run with more MPI ranks.

Thanks to Yang LI and Ludovic Metivier for reporting.