Fix in the Python interface for Gaussian processes.
By default ctypes assumes the return type of
a C function is an int (32 bit).

Our C/C++ code returns a pointer, which is 64 bit,
but only the lower 32 bits of this pointer were used.
This is fixed by explicitly specifying the return type using
  sp.STRUMPACK_create_general_kernel_double.restype = ctypes.c_void_p

Also when passing this pointer back to the C code,
it needs to be passed explicitly as a pointer.