Refactoring, move Front* in separate folder, split in hpp/cpp.
This revealed a serious numerical issue. In FrontalMatrixDense,
real_t was not defined. This means that real_t was taken from metis,
where it was defined as a float (not as the real part of scalar_t).

This might also be an issue in other places, will go check that now.
This can be prevented by splitting the (Par)Metis interface in a
separate source file.