Fixes in travis tests. Add tests for updating of matrix values. Change --sp_hss_min_sep_size to --sp_compression_min_sep_size Some tests failed with certain orderings, so I just picked a different ordering.
There was a change in the test code test_sparse_mpi. Before, MC64 and
symmetrization was performed before scattering the matrix over
all processes. Now MC64 and symmetrization are done in the parallel
solver. But this means the distribution of the matrix can be slightly
different, since it tries to give processors equal nonzeros. And this
can lead to different results for parmetis. THis is why some previously
working tests now fail.
The check for the error tolerance is probably not good.

This also adds tests for updating the matrix values and
reusing the permutation. The issue here is that MC64
should also be redone when the values change, but MC64 is done BEFORE
nested-dissection. So that should in theory also be redone.